requirements.txt
flask>=3.0
psycopg2-binary>=2.9
werkzeug>=3.0
python-dotenv>=1.0
